xkeyboard-config xserver pixman mesa libX11 xorgproto plink git update Jan 22 2019
xserver          commit 709c6562975c3bea10dd0571527a4aac79a6bf6f
xkeyboard-config commit fb188be4ccdf62ebaebe69fd1aacdefbaba8bcdc
libX11           commit 5538b3e4ae6dee32c47db9dfc85b07bbe7b90f6c
pixman           commit b40d5495ece2b2b5cff60e386c2f69db4ae03916
mesa             commit 27af1cc2a6c7d8f623f60d1499d2de2acf931534
putty            commit baff23cdd62ccc34a21534bf32b7779ff8853458
xorgproto        commit 000ebed576aafb44caeea8b6a5de90fba2bdc389

+++ libX11/src/XlibInt.c
@@ -50,6 +50,8 @@ from The Open Group.
 #ifdef XTHREADS
 #include "locking.h"
 
+#include <sys/ioctl.h>
+
 /* these pointers get initialized by XInitThreads */
 LockInfoPtr _Xglobal_lock = NULL;
 void (*_XCreateMutex_fn)(LockInfoPtr) = NULL;
@@ -1234,6 +1236,21 @@ _XWireToEvent(
 	return(True);
 }
 
+static int
+SocketBytesReadable(Display *dpy)
+{
+    int bytes = 0, last_error;
+#ifdef WIN32
+    last_error = WSAGetLastError();
+    ioctlsocket(ConnectionNumber(dpy), FIONREAD, &bytes);
+    WSASetLastError(last_error);
+#else
+    last_error = errno;
+    ioctl(ConnectionNumber(dpy), FIONREAD, &bytes);
+    errno = last_error;
+#endif
+    return bytes;
+}
 
 /*
  * _XDefaultIOError - Default fatal system error reporting routine.  Called
@@ -1242,25 +1259,35 @@ _XWireToEvent(
 _X_NORETURN int _XDefaultIOError(
 	Display *dpy)
 {
-	if (ECHECK(EPIPE)) {
-	    (void) fprintf (stderr,
-	"X connection to %s broken (explicit kill or server shutdown).\r\n",
-			    DisplayString (dpy));
-	} else {
-	    (void) fprintf (stderr,
-			"XIO:  fatal IO error %d (%s) on X server \"%s\"\r\n",
-#ifdef WIN32
-			WSAGetLastError(), strerror(WSAGetLastError()),
-#else
-			errno, strerror (errno),
-#endif
-			DisplayString (dpy));
-	    (void) fprintf (stderr,
-	 "      after %lu requests (%lu known processed) with %d events remaining.\r\n",
-			NextRequest(dpy) - 1, LastKnownRequestProcessed(dpy),
-			QLength(dpy));
+	int killed = ECHECK(EPIPE);
 
-	}
+	/*
+	 * If the socket was closed on the far end, the final recvmsg in
+	 * xcb will have thrown EAGAIN because we're non-blocking. Detect
+	 * this to get the more informative error message.
+	 */
+	if (ECHECK(EAGAIN) && SocketBytesReadable(dpy) <= 0)
+	    killed = True;
+
+	if (killed) {
+	    fprintf (stderr,
+                     "X connection to %s broken (explicit kill or server shutdown).\r\n",
+                     DisplayString (dpy));
+	} else {
+            fprintf (stderr,
+                     "XIO:  fatal IO error %d (%s) on X server \"%s\"\r\n",
+#ifdef WIN32
+                      WSAGetLastError(), strerror(WSAGetLastError()),
+#else
+                      errno, strerror (errno),
+#endif
+                      DisplayString (dpy));
+	    fprintf (stderr,
+		     "      after %lu requests (%lu known processed) with %d events remaining.\r\n",
+		     NextRequest(dpy) - 1, LastKnownRequestProcessed(dpy),
+		     QLength(dpy));
+        }
+
 	exit(1);
 	/*NOTREACHED*/
 }
